+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ### Version 3.14 [In Development]
+**All WoW Versions:**  
+ElvUI 13.83+ compatibility  
 **Retail:**  
 Updates for Shadow & Light profile (AFK Mode, Armory Mode)  
 ___